Fixes #48 Refactor Process
- Extract Cli
- Extract Config
- Encapsulate Shtats app and reduce arguments
- Update tests
- Move collector creation into process
