+++ tests/integration_test.rs
@@ -1,6 +1,7 @@
+use shtats::config::Config;
 use shtats::html::HtmlReporter;
 use shtats::output::BufferedOutput;
-use shtats::process::Config;
+use shtats::process::{Shtats};
 use shtats::Reporter;
 
 mod common;
@@ -11,11 +12,14 @@ fn test_cli_report() {
     common::setup_git_repo(|path| {
         let output = BufferedOutput::new();
         let mut reporter = HtmlReporter::new(output);
-        shtats::process::run_shtats(path, &mut reporter, Config{
+        let config = Config{
             until: None,
             since: None,
             output: "".to_string()
-        }, &|| {})?;
+        };
+        let process_callback =  &|| {};
+        let mut shtats = Shtats::create(&mut reporter, config, &process_callback);
+        shtats.run(path)?;
 
         assert!(reporter.to_string().contains("<tr>
                     <td>Number of commits_collection</td>
